address review comments: remove cast to an int in hashcode()

+++ capstone_project/src/main/java/com/google/sps/data/School.java
@@ -41,7 +41,7 @@ public class School {
 
   @Override
   public int hashCode() {
-    return (int) name.hashCode();
+    return name.hashCode();
   }
 
   @Override
